added users rol control:user, admin

+++ admin_post.php
@@ -1,5 +1,6 @@
 <?php
 include_once("model/db.php");
+include_once("model/user_obj.php");
 
 $sql = "SELECT post_id, title FROM post";
 $result = $db->fetch($sql);
@@ -16,20 +17,22 @@ while ($row = mysqli_fetch_array($result)) {
     <tr>
       <th> ID </th>
       <th> TITULO </th>
+    <?php if ($user->get_user_rol() == "admin") { ?>
       <th colspan="2"> OPERACIONES </th>
+    <?php } ?>
     </tr>
   </thead>
   <tbody>
-<?php
-foreach ($post as $key => $value) {
-  print '<tr>
-          <td>'. $key. '</td>
-          <td>'. $value['title']. '</td>
-          <td><a href="index.php?page=edit_post&post_id='. $key. '"> editar </a></td>
-          <td><a href="index.php?page=delete_post&post_id='. $key. '">  eliminar </a></td>
-        </tr>';
-}
-?>
+<?php foreach ($post as $key => $value) { ?>
+    <tr>
+      <td><?php print $key; ?></td>
+      <td><?php print $value['title']; ?></td>
+    <?php if ($user->get_user_rol() == "admin") { ?>
+      <td><a href="index.php?page=edit_post&post_id='. $key. '"> editar </a></td>
+      <td><a href="index.php?page=delete_post&post_id='. $key. '">  eliminar </a></td>
+    <?php } ?>
+    </tr>
+<?php } ?>
   </tbody>
 </table>
